# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "EduNFT API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:8050",
  "basePath": "/",
  "tags": [
    {
      "name": "Token",
      "description": "Token APIs"
    },
    {
      "name": "Image",
      "description": "Image APIs"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/Image/Upload": {
      "post": {
        "tags": [
          "Image"
        ],
        "description": "Upload image to IPFS",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "image",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Some description..."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Token/Mint": {
      "post": {
        "tags": [
          "Token"
        ],
        "description": "Mint NFT Token",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Token Information",
            "schema": {
              "$ref": "#/definitions/Mint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Token/Get/{id}": {
      "get": {
        "tags": [
          "Token"
        ],
        "description": "Mint NFT Token",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Image/Get/{cid}": {
      "get": {
        "tags": [
          "Image"
        ],
        "description": "Get Image Upload Info",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "fileName",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Mint": {
      "type": "object",
      "properties": {
        "diplomas": {
          "type": "object",
          "properties": {}
        },
        "images": {
          "type": "object",
          "properties": {}
        }
      },
      "required": [
        "diplomas",
        "images"
      ]
    }
  }
}